Fix packaging (six)
This fixes #683 six bug which is only another side-effect of the weird packaging implementation changes made to ansible-lint in its more recent history.

This change aims to address these issues by implementing proven working packaging practices, while removing custom hacks.

We will still need to do some follow-ups to remove implementation code from __init__.py which is a known source of problems. This patch avoided that in order to minimize changes.

Fixes: #683
