+++ pyproject.toml
@@ -4,13 +4,5 @@ requires = [
   "setuptools_scm >= 1.15.0",
   "setuptools_scm_git_archive >= 1.0",
   "wheel",
-
-  # needed because setuptools' attr import
-  # machinery (`importlib.import_module`)
-  # imports `__init__` first
-  "ansible",
-  "pathlib2; python_version < '3.2'",
-  "ruamel.yaml",
-  "six",
 ]
 build-backend = "setuptools.build_meta"
