+++ setup.cfg
@@ -21,7 +21,6 @@ ignore =
 
 [metadata]
 name = ansible-lint
-version = attr: ansiblelint.versiontools.get_self_version
 url = https://github.com/ansible/ansible-lint
 project_urls =
   Bug Tracker = https://github.com/ansible/ansible-lint/issues
