+++ setup.py
@@ -1,0 +1,8 @@
+#! /usr/bin/env python
+import setuptools
+
+
+if __name__ == "__main__":
+    setuptools.setup(
+        use_scm_version=True, setup_requires=["setuptools_scm"],
+    )
