+++ tox.ini
@@ -56,7 +56,8 @@ passenv =
 deps =
   pep517 >= 0.5.0
 commands =
-  {envpython} -c 'import os.path, shutil, sys; \
+  {envpython} -c 'from __future__ import print_function; \
+  import os.path, shutil, sys; \
   dist_dir = os.path.join("{toxinidir}", "dist"); \
   os.path.isdir(dist_dir) or sys.exit(0); \
   print("Removing \{!s\} contents...".format(dist_dir), file=sys.stderr); \
@@ -92,8 +93,10 @@ description =
   Verify that dists under the dist/ dir have valid metadata
 depends = build-dists
 deps =
+  {[testenv:build-dists]deps}
   twine
 skip_install = true
 # Ref: https://twitter.com/di_codes/status/1044358639081975813
 commands =
+  {[testenv:build-dists]commands}
   twine check {toxinidir}/dist/*
